requirements.txt
Django==5.2
Pillow==11.2.1
django-admin-sortable2==2.2.6
django-tinymce==4.1.0
django-environ==0.12.0